Rename pdp/ to product-detail/, move orphan files
- Rename pdp/ to product-detail/ for clarity
- Move banner-section.tsx to sections/
- Move error-boundary-content.tsx to ui/
