+++ apps/template/components/product-detail/about-item.tsx
