+++ apps/template/components/product-detail/auto-play-video.tsx
