+++ apps/template/components/product-detail/color-picker.tsx
