+++ apps/template/components/product-detail/lightbox.tsx
