+++ apps/template/components/product-detail/option-picker.tsx
