+++ apps/template/components/product-detail/product-info.tsx
