+++ apps/template/components/product-detail/product-media.tsx
