+++ apps/template/components/product-detail/product-price.tsx
