+++ apps/template/components/product-detail/schema.tsx
