+++ apps/template/components/product-detail/shop-logo.tsx
