+++ apps/template/components/sections/banner-section.tsx
