+++ apps/template/components/ui/error-boundary-content.tsx
